EAC3 works!
It does have the exact same seeking issue as FLAC. Darn! Now I need to actuallly try to fix that...

(Claude did that in one shot! Oh my god! In < 10 minutes it just went and gave QuickTime the ability to play EAC3 audio, all by itself. I mean, sure, we'd already laid the groundwork by making that FLAC codec which DID take a full day of back and forth, and the whole reason I did that was to lay the groundwork for more complicated codecs... but not THIS fast!)